backend/requirements.txt
# Core Web Framework
fastapi
uvicorn[standard]
websockets

# Supabase (Database)
supabase

# Google Gemini Live API
google-genai

# Audio processing
pyaudio

# Environment and Configuration
python-dotenv

# HTTP Client
aiohttp
requests

# Data processing
pydantic
numpy

# Database
asyncpg 